feat: Folder作成フィールドのレイアウト修正
